Add remainder_DGModel
Adds an outer constructor for `DGmodel` for remainder models. This will
allow the model to be a composition of `DGmodel`s as opposed to balance
laws. This will enable making the remainder model a true spitting in the
sense that the sum of all the models equals the original discretization.
This will also allow for more fine-grained control of the model
directionality (for instance, `EveryDirection()` for the main model with
`VerticalDirection()` acoustic model subtracted)